refactor: organize exceptions files
