+++ src/test/java/br/com/customers/application/usecases/customer/impl/FindAllCustomersUseCaseImplTest.java
@@ -1,7 +1,7 @@
 package br.com.customers.application.usecases.customer.impl;
 
 import br.com.customers.api.v1.model.PagedCustomerResponseDTO;
-import br.com.customers.application.exceptions.custom.InvalidParamException;
+import br.com.customers.application.exceptions.InvalidParamException;
 import br.com.customers.infrastructure.adapters.outbound.repositories.CustomerRepository;
 import br.com.customers.infrastructure.adapters.outbound.repositories.entities.CustomerEntity;
 import br.com.customers.infrastructure.mappers.CustomerMapper;
